updated README to conform to PRD in structure and formatting.
I removed the loadDictionaryFromFile as a public API, so I removed mention of this from the README file.

I added more context about cryptographic hash function properties, and why CrackStation is useful. Also a mission statement that discusses ethics.

Added installation instructions for both the Swift Package Manager and an Xcode app project.

Added a Usage section with the API function signatures and an example of usage.

Moved my Author section to the bottom.

removed the troubleshooting section, and combined the "note" section with the "Overview" section under a smaller heading. Also (hopefully) added more clarity in wording